+++ framework/db-core/src/main/java/is/codion/framework/db/AbstractEntityConnectionProvider.java
@@ -63,7 +63,7 @@ public abstract class AbstractEntityConnectionProvider implements EntityConnecti
 		this.user = requireNonNull(builder.user, "A user must be specified");
 		this.domain = requireNonNull(builder.domain, "A domain must be specified");
 		this.clientId = requireNonNull(builder.clientId, "A clientId must be specified");
-		this.clientType = builder.clientType;
+		this.clientType = builder.clientType == null ? domain.name() : builder.clientType;
 		this.clientVersion = builder.clientVersion;
 		this.onClose = builder.onClose;
 	}