+++ framework/domain-test/src/main/java/is/codion/framework/domain/test/DomainTest.java
@@ -89,7 +89,6 @@ public class DomainTest {
 	public DomainTest(Domain domain, Function<EntityConnection, EntityFactory> entityFactory, User user) {
 		this.connectionProvider = LocalEntityConnectionProvider.builder()
 						.domain(requireNonNull(domain))
-						.clientType(getClass().getName())
 						.user(requireNonNull(user))
 						.build();
 		this.entityFactory = requireNonNull(entityFactory);